+++ src/common/type.ts
@@ -2,7 +2,6 @@
 export type TempCache = {
   clipboard: string
   lastPage: string
-  visualMode: boolean
 };
 
 export type N = {